add 'update source' functionality

+++ src/app/source/source.component.ts
@@ -1,6 +1,7 @@
 import { Component, OnInit } from '@angular/core';
 import {HttpClient} from '@angular/common/http';
 import { CustomFormsModule } from 'ng2-validation';
+
 @Component({
   selector: 'app-source',
   templateUrl: './source.component.html',
@@ -17,10 +18,6 @@ export class SourceComponent implements OnInit {
   editItem: any;
 
   newSourceURL: string;
-  newTitle = 'title';
-  newDescription: 'description';
-  newLink: 'link';
-  newPubDate: 'pubDate';
 
   constructor(private httpClient: HttpClient) { }
 
@@ -29,10 +26,10 @@ export class SourceComponent implements OnInit {
 
     this.httpClient.get('http://localhost:8080/sources/getAll').subscribe((data) => {
       this.sourceItems = data;
-      this.title = 'title';
+      /*this.title = 'title';
       this.description = 'description';
       this.link = 'link';
-      this.pubDate = 'pubDate';
+      this.pubDate = 'pubDate';*/
     });
   }
 
@@ -40,25 +37,35 @@ export class SourceComponent implements OnInit {
 
   addSrc() {
     console.log('test');
-    const body = {
+
+    console.log(this.newSourceURL);
+    const regex = /^(?:http(s)?:\/\/)?[\w.-]+(?:\.[\w\.-]+)+[\w\-\._~:/?#[\]@!\$&'\(\)\*\+,;=.]+$/y;
+    const match = regex.exec(this.newSourceURL);
+    if (match == null) {
+      console.log('ttt');
+      (document.getElementById('src') as HTMLElement).style.borderBottom = '2px solid red';
+    } else {
+      const body = {
       sourceURL: this.newSourceURL,
-      title: this.title,
-      description: this.description,
-      link: this.link,
-      pub_date: this.pubDate,
+      title: document.getElementById('newTitle').value,
+      description: document.getElementById('newDescription').value,
+      link: document.getElementById('newLink').value,
+      pubDate: document.getElementById('newPubDate').value,
       };
-    this.httpClient.post('http://localhost:8080/sources/add', body).subscribe();
-
-    window.location.reload();
-  }
-
+      console.log(body);
+      this.httpClient.post('http://localhost:8080/sources/add', body).subscribe();
+      this.newSourceURL = '';
+      this.sourceItems.push(body);
+      // window.location.reload();
+    }
+}
 
   delete(id1: any) {
     console.log('delete');
-    //if (confirm('Are you sure you want to delete these source?')) {
+    if (confirm('Are you sure you want to delete these source?')) {
       this.httpClient.delete(`http://localhost:8080/sources/${id1}`).subscribe();
       window.location.reload();
-    //}
+    }
   }
 
   open(editItem: any) {
@@ -73,16 +80,17 @@ export class SourceComponent implements OnInit {
 
     const body = {
       id: ob.id,
-      sourceURL: this.sourceURL,
-      title: this.title,
-      description: this.description,
-      link: this.link,
+      sourceURL: document.getElementById('modifiedSourceURL').value,/*this.sourceURL,*/
+      title: document.getElementById('title').value,/*this.title,*/
+      description: document.getElementById('description').value,/*this.description,*/
+      link: document.getElementById('link').value,/*this.link,*/
       hostname: ob.hostname,
-      pubDate: this.pubDate
+      pubDate: document.getElementById('pubDate').value,/*this.pubDate,*/
     }
     console.log(body);
     this.httpClient.put(`http://localhost:8080/sources/${this.ids}`, body).subscribe();
-    window.location.reload();
+    
+    //window.location.reload();
 }
